+++ src/use-cases/general-behaviors/entity-media/set-entity-image/0006-confirm-and-persist/external-url.stories.tsx
@@ -143,7 +143,9 @@ function ExternalUrlScene({ sceneIndex }: { sceneIndex: number }) {
           <ImagePreviewEditor
             aspectRatio={1}
             imageData={MOCK_ITEM_IMAGE}
-            onCropChange={noop}
+            onCropComplete={noop}
+            onZoomChange={noop}
+            onRotationChange={noop}
             onReset={noop}
           />
           <div className="mt-4">